- 로그인 처리 - 로그인 후 추가, 수정, 삭제 처리 - navbar, button 등 로그인 사용자만 출력되게 처리 완료.

+++ src/main/java/com/prj1/controller/BoardController.java
@@ -4,14 +4,14 @@ import com.prj1.domain.Board;
 import com.prj1.service.BoardService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-
-import java.util.List;
 
 @Controller
 @RequiredArgsConstructor
@@ -26,10 +26,9 @@ public class BoardController {
     }
 
     @PostMapping("/add")
-    public String add(@ModelAttribute Board board, RedirectAttributes rttr) {
-        log.info("Board={}", board);
+    public String add(@ModelAttribute Board board, Authentication authentication, RedirectAttributes rttr) {
+        service.add(board, authentication);
 
-        service.add(board);
         rttr.addAttribute("id", board.getId());
         return "redirect:/board";
     }
@@ -43,16 +42,18 @@ public class BoardController {
     }
 
     @GetMapping("/")
-    public String home(Model model) {
-        List<Board> list = service.list();
-        model.addAttribute("boardList", list);
+    public String home(@RequestParam(defaultValue = "1") Integer page, Model model) {
+
+        model.addAllAttributes(service.list(page));
         return "board/home";
     }
 
     @PostMapping("/delete")
-    public String delete(Integer id, RedirectAttributes rttr) {
-        int row = service.delete(id);
-        rttr.addFlashAttribute("delete", row);
+    public String delete(Integer id, Authentication authentication, RedirectAttributes rttr) {
+        if (service.hasAccess(id, authentication)) {
+            service.delete(id);
+        }
+
         rttr.addFlashAttribute("deleteId", id);
         return "redirect:/";
     }
@@ -65,8 +66,11 @@ public class BoardController {
     }
 
     @PostMapping("/update")
-    public String update(Board board, RedirectAttributes rttr) {
-        service.update(board);
+    public String update(Board board, Authentication authentication, RedirectAttributes rttr) {
+
+        if (service.hasAccess(board.getId(), authentication)) {
+            service.update(board);
+        }
         rttr.addAttribute("id", board.getId());
         return "redirect:/board";
     }
